Yet more cleanup, mostly minor style.
Replace manual HTML escaping with toHtmlEscaped().
Replace old-style connect() with the function pointer version.
Replace calls to QString(const char*) with QStringLiteral() in a bunch of places.
Replace a few instances of 0 with nullptr.
Remove a couple of dead/redundant slots.
Change much of the indenting in flist_messenger.cpp to be less spacey and more tabby.